requirements.txt
Django==2.2.1
django-crispy-forms==1.8.0
djangorestframework==3.10.3
Pillow==6.2.1
pytz==2019.1
sqlparse==0.3.0